login-session

+++ PredragRadak_Projekat_2021_web/main.py
@@ -1,9 +1,7 @@
 import flask
 from flask import Flask
 
-# import korisnik
-# import kupac
-# import knjiga
+from flask import session
 
 from flaskext.mysql import MySQL
 from flaskext.mysql import pymysql
@@ -22,15 +20,109 @@ app.config["MYSQL_DATABASE_HOST"] = "localhost"
 app.config["MYSQL_DATABASE_USER"] = "root"
 app.config["MYSQL_DATABASE_PASSWORD"] = "pedja10"
 app.config["MYSQL_DATABASE_DB"] = "prodaja_knjiga"
+app.secret_key = "PR.12.rp.sr/mr"
 
 
 mysql = MySQL(app, cursorclass=pymysql.cursors.DictCursor)
-
 
 #######Pocetna strana
 @app.route("/")
 def home():
     return app.send_static_file("index.html")
+
+#####LOGIN####
+@app.route("/api/login", methods=["POST"]) #Novo-Logovanje korisnika
+def login():
+    cursor = mysql.get_db().cursor()
+    cursor.execute("SELECT * FROM korisnik WHERE email=%(email)s AND lozinka=%(lozinka)s", flask.request.json)
+    korisnik = cursor.fetchone()
+    if korisnik is not None:
+        session["korisnik"] = korisnik["email"]
+        return "", 200
+    return "", 403
+
+@app.route("/api/logout", methods=["GET"]) ###Novo-Odjava
+def logout():
+    session.pop("korisnik", None)
+    return "", 200
+#################
+
+#####LOGIN####
+@app.route("/api/loginBibliotekar", methods=["POST"]) #Novo-Logovanje bibliotekara
+def loginBibliotekar():
+    cursor = mysql.get_db().cursor()
+    cursor.execute("SELECT * FROM bibliotekar WHERE korisnicko_ime=%(korisnicko_ime)s AND lozinka=%(lozinka)s", flask.request.json)
+    bibliotekar = cursor.fetchone()
+    if bibliotekar is not None:
+        session["bibliotekar"] = bibliotekar["korisnicko_ime"]
+        return "", 200
+    return "", 403
+
+@app.route("/api/logoutBibliotekar", methods=["GET"]) ###Novo-Odjava
+def logoutBibliotekar():
+    session.pop("bibliotekar", None)
+    return "", 200
+#################
+
+#Bibliotekar
+@app.route("/api/bibliotekari")
+def getAllBibliotekar():
+    if session.get("bibliotekar") is not None:
+        cursor = mysql.get_db().cursor()
+        cursor.execute("SELECT * FROM bibliotekar")
+        bibliotekar = cursor.fetchall()
+        return flask.jsonify(bibliotekar)
+    return "", 401   #Login return od ifa
+########Prikaz liste###########
+@app.route("/api/bibliotekari/<int:bibliotekar_id>")
+def getBibliotekar(bibliotekar_id):
+    if session.get("bibliotekar") is not None:
+        cursor = mysql.get_db().cursor()
+        cursor.execute("SELECT * FROM bibliotekar WHERE id=%s", (bibliotekar_id,))
+        bibliotekar=cursor.fetchone()
+        if bibliotekar is not None:
+            return flask.jsonify(bibliotekar)
+        return "", 404
+    return "", 401   #Login return od ifa
+# #################################
+@app.route("/api/bibliotekari", methods=["POST"])
+def dodajBibliotekara():
+    db = mysql.get_db()
+    cursor = db.cursor()
+    cursor.execute("INSERT INTO bibliotekar(korisnicko_ime, lozinka) VALUES (%(korisnicko_ime)s, %(lozinka)s)", flask.request.json)
+    db.commit()
+    return flask.request.json, 201
+    
+
+@app.route("/api/bibliotekari/<int:bibliotekar_id>", methods=["DELETE"])
+def ukloniBibliotekara(bibliotekar_id):
+    if session.get("bibliotekar") is not None:
+        db = mysql.get_db()
+        cursor = db.cursor()
+        cursor.execute("DELETE FROM bibliotekar WHERE id=%s", (bibliotekar_id, ))
+        db.commit()
+        return ""
+    return "", 401   #Login return od ifa
+
+@app.route("/api/bibliotekari/<int:bibliotekar_id>", methods=["PUT"])
+def izmeniBibliotekara(bibliotekar_id):
+    if session.get("bibliotekar") is not None:
+        bibliotekar = dict(flask.request.json)
+        bibliotekar["bibliotekar_id"] = bibliotekar_id
+        db = mysql.get_db()
+        cursor = db.cursor()
+        cursor.execute("UPDATE bibliotekar SET korisnicko_ime=%(korisnicko_ime)s, lozinka=%(lozinka)s WHERE id=%(bibliotekar_id)s", bibliotekar)
+        db.commit()
+        cursor.execute("SELECT * FROM bibliotekar WHERE id=%s", (bibliotekar_id, ))
+        bibliotekar = cursor.fetchone()
+        return flask.jsonify(bibliotekar)
+    return "", 401   #Login return od ifa
+
+
+
+
+
+
 
 #KORISNIK
 @app.route("/api/korisnici")
@@ -53,7 +145,7 @@ def getKorisnik(korisnik_IDKorisnik):
 def dodajKorisnika():
     db = mysql.get_db()
     cursor = db.cursor()
-    cursor.execute("INSERT INTO korisnik(ime, email, lozinka) VALUES (%(ime)s, %(email)s, %(lozinka)s)", flask.request.json)
+    cursor.execute("INSERT INTO korisnik(email, lozinka) VALUES (%(email)s, %(lozinka)s)", flask.request.json)
     db.commit()
     return flask.request.json, 201
 
@@ -71,7 +163,7 @@ def izmeniKorisnika(korisnik_IDKorisnik):
     korisnik["korisnik_IDKorisnik"] = korisnik_IDKorisnik
     db = mysql.get_db()
     cursor = db.cursor()
-    cursor.execute("UPDATE korisnik SET ime=%(ime)s, email=%(email)s, lozinka=%(lozinka)s WHERE IDKorisnik=%(korisnik_IDKorisnik)s", korisnik)
+    cursor.execute("UPDATE korisnik SET email=%(email)s, lozinka=%(lozinka)s WHERE IDKorisnik=%(korisnik_IDKorisnik)s", korisnik)
     db.commit()
     cursor.execute("SELECT * FROM korisnik WHERE IDKorisnik=%s", (korisnik_IDKorisnik, ))
     korisnik = cursor.fetchone()
@@ -81,10 +173,12 @@ def izmeniKorisnika(korisnik_IDKorisnik):
 ####Kupac
 @app.route("/api/kupci")
 def getAllKupci():
-    cursor = mysql.get_db().cursor()
-    cursor.execute("SELECT * FROM kupac")
-    kupac = cursor.fetchall()
-    return flask.jsonify(kupac)
+    if session.get("korisnik") is not None:  #Login za prikaz kupaca (Korisnik je povezan trenutno samo za kupce)
+        cursor = mysql.get_db().cursor()       ##Ovako sa if session mozes sve entitete
+        cursor.execute("SELECT * FROM kupac")
+        kupac = cursor.fetchall()
+        return flask.jsonify(kupac)
+    return "", 401   #Login return od ifa
 ########Prikaz liste###########
 @app.route("/api/kupci/<int:kupac_IDKupac>")
 def getKupac(kupac_IDKupac):
@@ -127,12 +221,15 @@ def izmeniKupca(kupac_IDKupac):
 ##########Knjiga
 @app.route("/api/knjige")
 def getAllKnjige():
-    cursor = mysql.get_db().cursor()
-    cursor.execute("SELECT * FROM knjiga")
-    knjiga = cursor.fetchall()
-    for k in knjiga:
-        k["cena"] = float(k["cena"])
-    return flask.jsonify(knjiga)
+    if session.get("korisnik") is not None:  #Login za prikaz kupaca (Korisnik je povezan trenutno samo za kupce)
+        cursor = mysql.get_db().cursor()
+        cursor.execute("SELECT * FROM knjiga")
+        knjiga = cursor.fetchall()
+        for k in knjiga:
+            k["cena"] = float(k["cena"])
+        return flask.jsonify(knjiga)
+    return "", 401   #Login return od ifa
+
  ########Prikaz liste###########
 @app.route("/api/knjige/<int:knjiga_IDKnjiga>")
 def getKnjiga(knjiga_IDKnjiga):
@@ -146,32 +243,38 @@ def getKnjiga(knjiga_IDKnjiga):
 #################################
 @app.route("/api/knjige", methods=["POST"])
 def dodajKnjige():
-    db = mysql.get_db()
-    cursor = db.cursor()
-    cursor.execute("INSERT INTO knjiga(naziv, autor, kategorija, cena, stanje, biblioteka_id) VALUES (%(naziv)s, %(autor)s, %(kategorija)s, %(cena)s, %(stanje)s, %(biblioteka_id)s)", flask.request.json)
-    db.commit()
-    return flask.request.json, 201
+    if session.get("bibliotekar") is not None:
+        db = mysql.get_db()
+        cursor = db.cursor()
+        cursor.execute("INSERT INTO knjiga(naziv, autor, kategorija, cena, stanje, biblioteka_id) VALUES (%(naziv)s, %(autor)s, %(kategorija)s, %(cena)s, %(stanje)s, %(biblioteka_id)s)", flask.request.json)
+        db.commit()
+        return flask.request.json, 201
+    return "", 401   #Login return od ifa   
 
 @app.route("/api/knjige/<int:knjiga_IDKnjiga>", methods=["DELETE"])
 def ukloniKnjigu(knjiga_IDKnjiga):
-    db = mysql.get_db()
-    cursor = db.cursor()
-    cursor.execute("DELETE FROM knjiga WHERE IDKnjiga=%s", (knjiga_IDKnjiga, ))
-    db.commit()
-    return ""
+    if session.get("bibliotekar") is not None:
+        db = mysql.get_db()
+        cursor = db.cursor()
+        cursor.execute("DELETE FROM knjiga WHERE IDKnjiga=%s", (knjiga_IDKnjiga, ))
+        db.commit()
+        return ""
+    return "", 401   #Login return od ifa   
 
 @app.route("/api/knjige/<int:knjiga_IDKnjiga>", methods=["PUT"])
 def izmeniKnjigu(knjiga_IDKnjiga):
-    knjiga = dict(flask.request.json)
-    knjiga["knjiga_IDKnjiga"] = knjiga_IDKnjiga
-    db = mysql.get_db()
-    cursor = db.cursor()
-    cursor.execute("UPDATE knjiga SET naziv=%(naziv)s, autor=%(autor)s, kategorija=%(kategorija)s, cena=%(cena)s, stanje=%(stanje)s, biblioteka_id=%(biblioteka_id)s WHERE IDKnjiga=%(knjiga_IDKnjiga)s", knjiga)
-    db.commit()
-    cursor.execute("SELECT * FROM knjiga WHERE IDKnjiga=%s", (knjiga_IDKnjiga, ))
-    knjiga = cursor.fetchone()
-    knjiga["cena"] = float(knjiga["cena"]) ###Decimal u bazi!!!
-    return flask.jsonify(knjiga)
+    if session.get("bibliotekar") is not None:
+        knjiga = dict(flask.request.json)
+        knjiga["knjiga_IDKnjiga"] = knjiga_IDKnjiga
+        db = mysql.get_db()
+        cursor = db.cursor()
+        cursor.execute("UPDATE knjiga SET naziv=%(naziv)s, autor=%(autor)s, kategorija=%(kategorija)s, cena=%(cena)s, stanje=%(stanje)s, biblioteka_id=%(biblioteka_id)s WHERE IDKnjiga=%(knjiga_IDKnjiga)s", knjiga)
+        db.commit()
+        cursor.execute("SELECT * FROM knjiga WHERE IDKnjiga=%s", (knjiga_IDKnjiga, ))
+        knjiga = cursor.fetchone()
+        knjiga["cena"] = float(knjiga["cena"]) ###Decimal u bazi!!!
+        return flask.jsonify(knjiga)
+    return "", 401   #Login return od ifa   
 
 ####Biblioteka
 @app.route("/api/biblioteke")
